% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountPermission}
\alias{AccountPermission}
\title{AccountPermission Object}
\usage{
AccountPermission(accountProfiles = NULL, id = NULL, level = NULL,
  name = NULL, permissionGroupId = NULL)
}
\arguments{
\item{accountProfiles}{Account profiles associated with this account permission}

\item{id}{ID of this account permission}

\item{level}{Administrative level required to enable this account permission}

\item{name}{Name of this account permission}

\item{permissionGroupId}{Permission group of this account permission}
}
\value{
AccountPermission object
}
\description{
AccountPermission Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AccountPermissions contains information about a particular account permission. Some features of DCM require an account permission to be present in the account.
}

