% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{AccountPermissionGroupsListResponse}
\alias{AccountPermissionGroupsListResponse}
\title{AccountPermissionGroupsListResponse Object}
\usage{
AccountPermissionGroupsListResponse(accountPermissionGroups = NULL)
}
\arguments{
\item{accountPermissionGroups}{Account permission group collection}
}
\value{
AccountPermissionGroupsListResponse object
}
\description{
AccountPermissionGroupsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Account Permission Group List Response
}

