% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{BrowsersListResponse}
\alias{BrowsersListResponse}
\title{BrowsersListResponse Object}
\usage{
BrowsersListResponse(browsers = NULL)
}
\arguments{
\item{browsers}{Browser collection}
}
\value{
BrowsersListResponse object
}
\description{
BrowsersListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Browser List Response
}

