% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CitiesListResponse}
\alias{CitiesListResponse}
\title{CitiesListResponse Object}
\usage{
CitiesListResponse(cities = NULL)
}
\arguments{
\item{cities}{City collection}
}
\value{
CitiesListResponse object
}
\description{
CitiesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
City List Response
}

