% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ConnectionTypesListResponse}
\alias{ConnectionTypesListResponse}
\title{ConnectionTypesListResponse Object}
\usage{
ConnectionTypesListResponse(connectionTypes = NULL)
}
\arguments{
\item{connectionTypes}{Collection of connection types such as broadband and mobile}
}
\value{
ConnectionTypesListResponse object
}
\description{
ConnectionTypesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Connection Type List Response
}

