% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Conversion}
\alias{Conversion}
\title{Conversion Object}
\usage{
Conversion(childDirectedTreatment = NULL, customVariables = NULL,
  encryptedUserId = NULL, floodlightActivityId = NULL,
  floodlightConfigurationId = NULL, limitAdTracking = NULL,
  mobileDeviceId = NULL, ordinal = NULL, quantity = NULL,
  timestampMicros = NULL, value = NULL)
}
\arguments{
\item{childDirectedTreatment}{Whether the conversion was directed toward children}

\item{customVariables}{Custom floodlight variables}

\item{encryptedUserId}{The alphanumeric encrypted user ID}

\item{floodlightActivityId}{Floodlight Activity ID of this conversion}

\item{floodlightConfigurationId}{Floodlight Configuration ID of this conversion}

\item{limitAdTracking}{Whether the user has Limit Ad Tracking set}

\item{mobileDeviceId}{The mobile device ID}

\item{ordinal}{The ordinal of the conversion}

\item{quantity}{The quantity of the conversion}

\item{timestampMicros}{The timestamp of conversion, in Unix epoch micros}

\item{value}{The value of the conversion}
}
\value{
Conversion object
}
\description{
Conversion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Conversion represents when a user successfully performs a desired action after seeing an ad.
}

