% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ConversionsBatchInsertResponse}
\alias{ConversionsBatchInsertResponse}
\title{ConversionsBatchInsertResponse Object}
\usage{
ConversionsBatchInsertResponse(hasFailures = NULL, status = NULL)
}
\arguments{
\item{hasFailures}{Indicates that some or all conversions failed to insert}

\item{status}{The status of each conversion's insertion status}
}
\value{
ConversionsBatchInsertResponse object
}
\description{
ConversionsBatchInsertResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Insert Conversions Response.
}

