% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{creativeFields.patch}
\alias{creativeFields.patch}
\title{Updates an existing creative field. This method supports patch semantics.}
\usage{
creativeFields.patch(CreativeField, profileId, id)
}
\arguments{
\item{CreativeField}{The \link{CreativeField} object to pass to this method}

\item{profileId}{User profile ID associated with this request}

\item{id}{Creative Field ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other CreativeField functions: \code{\link{CreativeField}},
  \code{\link{creativeFields.insert}},
  \code{\link{creativeFields.update}}
}

