% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{CustomRichMediaEvents}
\alias{CustomRichMediaEvents}
\title{CustomRichMediaEvents Object}
\usage{
CustomRichMediaEvents(filteredEventIds = NULL)
}
\arguments{
\item{filteredEventIds}{List of custom rich media event IDs}
}
\value{
CustomRichMediaEvents object
}
\description{
CustomRichMediaEvents Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Custom Rich Media Events group.
}

