% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DirectorySiteContactAssignment}
\alias{DirectorySiteContactAssignment}
\title{DirectorySiteContactAssignment Object}
\usage{
DirectorySiteContactAssignment(contactId = NULL, visibility = NULL)
}
\arguments{
\item{contactId}{ID of this directory site contact}

\item{visibility}{Visibility of this directory site contact assignment}
}
\value{
DirectorySiteContactAssignment object
}
\description{
DirectorySiteContactAssignment Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Directory Site Contact Assignment
}

