% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{DynamicTargetingKey}
\alias{DynamicTargetingKey}
\title{DynamicTargetingKey Object}
\usage{
DynamicTargetingKey(name = NULL, objectId = NULL, objectType = NULL)
}
\arguments{
\item{name}{Name of this dynamic targeting key}

\item{objectId}{ID of the object of this dynamic targeting key}

\item{objectType}{Type of the object of this dynamic targeting key}
}
\value{
DynamicTargetingKey object
}
\description{
DynamicTargetingKey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of a dynamic targeting key. Dynamic targeting keys are unique, user-friendly labels, created at the advertiser level in DCM, that can be assigned to ads, creatives, and placements and used for targeting with DoubleClick Studio dynamic creatives. Use these labels instead of numeric DCM IDs (such as placement IDs) to save time and avoid errors in your dynamic feeds.
}
\seealso{
Other DynamicTargetingKey functions: \code{\link{dynamicTargetingKeys.insert}}
}

