% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{EventTagsListResponse}
\alias{EventTagsListResponse}
\title{EventTagsListResponse Object}
\usage{
EventTagsListResponse(eventTags = NULL)
}
\arguments{
\item{eventTags}{Event tag collection}
}
\value{
EventTagsListResponse object
}
\description{
EventTagsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Event Tag List Response
}

