% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FileList}
\alias{FileList}
\title{FileList Object}
\usage{
FileList(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{The eTag of this response for caching purposes}

\item{items}{The files returned in this response}

\item{nextPageToken}{Continuation token used to page through files}
}
\value{
FileList object
}
\description{
FileList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the list of File resources.
}

