% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{floodlightActivities.insert}
\alias{floodlightActivities.insert}
\title{Inserts a new floodlight activity.}
\usage{
floodlightActivities.insert(FloodlightActivity, profileId)
}
\arguments{
\item{FloodlightActivity}{The \link{FloodlightActivity} object to pass to this method}

\item{profileId}{User profile ID associated with this request}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}

Other FloodlightActivity functions: \code{\link{FloodlightActivity}},
  \code{\link{floodlightActivities.patch}},
  \code{\link{floodlightActivities.update}}
}

