% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{FrequencyCap}
\alias{FrequencyCap}
\title{FrequencyCap Object}
\usage{
FrequencyCap(duration = NULL, impressions = NULL)
}
\arguments{
\item{duration}{Duration of time, in seconds, for this frequency cap}

\item{impressions}{Number of times an individual user can be served the ad within the specified duration}
}
\value{
FrequencyCap object
}
\description{
FrequencyCap Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Frequency Cap.
}

