% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{LandingPagesListResponse}
\alias{LandingPagesListResponse}
\title{LandingPagesListResponse Object}
\usage{
LandingPagesListResponse(landingPages = NULL)
}
\arguments{
\item{landingPages}{Landing page collection}
}
\value{
LandingPagesListResponse object
}
\description{
LandingPagesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Landing Page List Response
}

