% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementsListResponse}
\alias{PlacementsListResponse}
\title{PlacementsListResponse Object}
\usage{
PlacementsListResponse(nextPageToken = NULL, placements = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{placements}{Placement collection}
}
\value{
PlacementsListResponse object
}
\description{
PlacementsListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Placement List Response
}

