% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{PlacementStrategiesListResponse}
\alias{PlacementStrategiesListResponse}
\title{PlacementStrategiesListResponse Object}
\usage{
PlacementStrategiesListResponse(nextPageToken = NULL,
  placementStrategies = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token to be used for the next list operation}

\item{placementStrategies}{Placement strategy collection}
}
\value{
PlacementStrategiesListResponse object
}
\description{
PlacementStrategiesListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Placement Strategy List Response
}

