% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{Pricing}
\alias{Pricing}
\title{Pricing Object}
\usage{
Pricing(capCostType = NULL, endDate = NULL, flights = NULL,
  groupType = NULL, pricingType = NULL, startDate = NULL)
}
\arguments{
\item{capCostType}{Cap cost type of this inventory item}

\item{endDate}{End date of this inventory item}

\item{flights}{Flights of this inventory item}

\item{groupType}{Group type of this inventory item if it represents a placement group}

\item{pricingType}{Pricing type of this inventory item}

\item{startDate}{Start date of this inventory item}
}
\value{
Pricing object
}
\description{
Pricing Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Pricing Information
}

