% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_functions.R
\name{remarketingLists.list}
\alias{remarketingLists.list}
\title{Retrieves a list of remarketing lists, possibly filtered.}
\usage{
remarketingLists.list(profileId, advertiserId, active = NULL,
  floodlightActivityId = NULL, maxResults = NULL, name = NULL,
  pageToken = NULL, sortField = NULL, sortOrder = NULL)
}
\arguments{
\item{profileId}{User profile ID associated with this request}

\item{advertiserId}{Select only remarketing lists owned by this advertiser}

\item{active}{Select only active or only inactive remarketing lists}

\item{floodlightActivityId}{Select only remarketing lists that have this floodlight activity ID}

\item{maxResults}{Maximum number of results to return}

\item{name}{Allows searching for objects by name or ID}

\item{pageToken}{Value of the nextPageToken from the previous result page}

\item{sortField}{Field by which to sort the list}

\item{sortOrder}{Order of sorted results, default is ASCENDING}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/dfatrafficking
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/dfatrafficking)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/doubleclick-advertisers/reporting/}{Google Documentation}
}

