% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{ThirdPartyTrackingUrl}
\alias{ThirdPartyTrackingUrl}
\title{ThirdPartyTrackingUrl Object}
\usage{
ThirdPartyTrackingUrl(thirdPartyUrlType = NULL, url = NULL)
}
\arguments{
\item{thirdPartyUrlType}{Third-party URL type for in-stream video creatives}

\item{url}{URL for the specified third-party URL type}
}
\value{
ThirdPartyTrackingUrl object
}
\description{
ThirdPartyTrackingUrl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Third-party Tracking URL.
}

