% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserRole}
\alias{UserRole}
\title{UserRole Object}
\usage{
UserRole(accountId = NULL, defaultUserRole = NULL, id = NULL,
  name = NULL, parentUserRoleId = NULL, permissions = NULL,
  subaccountId = NULL)
}
\arguments{
\item{accountId}{Account ID of this user role}

\item{defaultUserRole}{Whether this is a default user role}

\item{id}{ID of this user role}

\item{name}{Name of this user role}

\item{parentUserRoleId}{ID of the user role that this user role is based on or copied from}

\item{permissions}{List of permissions associated with this user role}

\item{subaccountId}{Subaccount ID of this user role}
}
\value{
UserRole object
}
\description{
UserRole Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties of auser role, which is used to manage user access.
}
\seealso{
Other UserRole functions: \code{\link{userRoles.insert}},
  \code{\link{userRoles.patch}},
  \code{\link{userRoles.update}}
}

