% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfareporting_objects.R
\name{UserRolePermissionGroup}
\alias{UserRolePermissionGroup}
\title{UserRolePermissionGroup Object}
\usage{
UserRolePermissionGroup(id = NULL, name = NULL)
}
\arguments{
\item{id}{ID of this user role permission}

\item{name}{Name of this user role permission group}
}
\value{
UserRolePermissionGroup object
}
\description{
UserRolePermissionGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a grouping of related user role permissions.
}

